sw: Move `datagen.py` and `verify.py` scripts to `scripts/`
