+++ sw/blas/axpy/Makefile
@@ -16,7 +16,7 @@ APP     ?= axpy
 SRCS    ?= $(SRC_DIR)/main.c
 INCDIRS += $(DATA_DIR) $(SRC_DIR)
 
-DATAGEN_PY = $(DATA_DIR)/datagen.py
+DATAGEN_PY = $(MK_DIR)/scripts/datagen.py
 DATA_H     = $(DATA_DIR)/data.h
 
 $(DATA_H): $(DATAGEN_PY)
