+++ sw/blas/axpy/scripts/datagen.py
