+++ sw/blas/axpy/scripts/verify.py
@@ -7,9 +7,9 @@
 
 import sys
 from pathlib import Path
-from data.datagen import golden_model
+from datagen import golden_model
 
-sys.path.append(str(Path(__file__).parent / '../../../util/sim/'))
+sys.path.append(str(Path(__file__).parent / '../../../../util/sim/'))
 from verif_utils import Verifier  # noqa: E402
