+++ sw/dnn/batchnorm/scripts/datagen.py
