+++ sw/dnn/conv2d/scripts/datagen.py
