+++ sw/dnn/flashattention_2/scripts/datagen.py
