+++ sw/dnn/fusedconv/scripts/datagen.py
