+++ sw/dnn/gelu/scripts/datagen.py
