+++ sw/dnn/layernorm/scripts/datagen.py
