+++ sw/dnn/maxpool/scripts/datagen.py
