+++ sw/dnn/softmax/scripts/datagen.py
